+++ src/core/config/mod.rs
@@ -1,3 +1,4 @@
+mod cache_size;
 mod config_msg;
 mod config_struct;
 mod json;
@@ -9,7 +10,7 @@ mod song_pos_save;
 
 use std::path::PathBuf;
 
-pub use self::{config_msg::*, config_struct::*};
+pub use self::{cache_size::*, config_msg::*, config_struct::*};
 
 /// Unique app identifier, it is different when debugging.
 #[cfg(not(debug_assertions))]
@@ -26,19 +27,19 @@ pub const VERSION_STR: &str = {
 
 /// Gets the default path for configuration, it is different when debugging.
 pub fn default_config_dir() -> PathBuf {
-    if let Some(dir) = dirs::config_dir() {
-        dir.join(APP_ID)
-    } else {
-        PathBuf::from(".")
-    }
+    get_uamp_dir(dirs::config_dir())
 }
 
 /// Gets the default path for logs.
 pub fn default_log_dir() -> PathBuf {
-    let mut d = dirs::data_local_dir()
-        .map(|a| a.join(APP_ID))
-        .unwrap_or_else(|| ".".into());
+    let mut d = get_uamp_dir(dirs::data_local_dir());
     d.push("log");
+    d
+}
+
+pub fn default_cache_dir() -> PathBuf {
+    let mut d = get_uamp_dir(dirs::cache_dir());
+    d.push("cache");
     d
 }
 
@@ -53,3 +54,11 @@ pub const DEBUG_PORT: u16 = 33284;
 pub const DEFAULT_PORT: u16 = RELEASE_PORT;
 #[cfg(debug_assertions)]
 pub const DEFAULT_PORT: u16 = DEBUG_PORT;
+
+fn get_uamp_dir(base: Option<PathBuf>) -> PathBuf {
+    base.map(|mut a| {
+        a.push(APP_ID);
+        a
+    })
+    .unwrap_or_else(|| ".".into())
+}